backend/requirements.txt
fastapi
python-multipart
pillow
opencv-python
numpy
pytesseract
python-dotenv
transformers
torch
sentencepiece 